fs/devfs: Use common devfs_iterate()
